app/requirements.txt
fastapi[all]
pydantic
uvicorn
scikit-learn
pandas
joblib
numpy
python-dotenv